requirements.txt
msgpack-python
coverage
pyyaml
